requirements.txt
# ETL Pipeline Dependencies
# Python 3.12

# Configuration
python-dotenv>=1.0.0

# Storage (S3-compatible)
boto3>=1.34.0
botocore>=1.34.0

# HTTP Requests
requests>=2.31.0
beautifulsoup4>=4.12.0

# Web Scraping (Playwright)
playwright>=1.40.0
playwright-stealth>=1.0.6

# Database
psycopg2-binary>=2.9.9

# Utilities
python-dateutil>=2.8.2
